requirements.txt
# Paste to requirements.txt
flask
google-generativeai
pillow
PyMuPDF
python-docx
importlib-metadata